+++ crates/nostr/src/nips/mod.rs
@@ -14,8 +14,6 @@ pub mod nip04;
 pub mod nip05;
 #[cfg(feature = "nip06")]
 pub mod nip06;
-#[cfg(all(feature = "nip07", target_arch = "wasm32"))]
-pub mod nip07;
 pub mod nip09;
 pub mod nip10;
 #[cfg(all(feature = "std", feature = "nip11"))]